Restructured The Component Dialog of the alarm tree plus cleanup
