--- app/alarm/ui/src/main/resources/org/phoebus/applications/alarm/ui/config/ComponentConfigDialog.fxml
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import org.phoebus.applications.alarm.ui.config.*?>
-
-<ScrollPane fx:id="scroll" fitToWidth="true" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.phoebus.applications.alarm.ui.config.ComponentConfigDialogController">
-
-    <GridPane fx:id="layout" hgap="5" vgap="5">
-        <padding>
-            <Insets bottom="5" left="5" right="5" top="5" />
-        </padding>
-
-        <columnConstraints>
-            <!-- Column 0: fixed 190 px label column -->
-            <ColumnConstraints maxWidth="190" minWidth="190" prefWidth="190" />
-            <!-- Column 1: grows to fill available width -->
-            <ColumnConstraints hgrow="ALWAYS" />
-        </columnConstraints>
-
-        <!-- ═══════════════════════════════════════════
-             Row 0 – Path  (read-only)
-             ═══════════════════════════════════════════ -->
-        <Label text="%path" GridPane.columnIndex="0" GridPane.rowIndex="0" />
-        <TextField fx:id="path" editable="false" GridPane.columnIndex="1" GridPane.hgrow="ALWAYS" GridPane.rowIndex="0" />
-
-        <!-- Row 1 – Behavior checkboxes -->
-        <Label fx:id="behaviorLabel" text="%behavior" GridPane.columnIndex="0" GridPane.rowIndex="1" />
-        <HBox alignment="CENTER_LEFT" spacing="10" GridPane.columnIndex="1" GridPane.rowIndex="1">
-            <CheckBox fx:id="enabled" mnemonicParsing="false" text="%enabled" />
-            <Label fx:id="partlyDisabledLabel" mnemonicParsing="false" text="%partlyDisabled2" visible="false"/>
-         <GridPane.margin>
-            <Insets bottom="5.0" />
-         </GridPane.margin>
-        </HBox>
-
-        <!-- Row 2 – Disable until (DateTimePicker + relative ComboBox) -->
-        <Label fx:id="disableUntilLabel" text="%disableUntil" GridPane.columnIndex="0" GridPane.rowIndex="2" />
-
-        <HBox fx:id="untilBox" alignment="CENTER" spacing="10" GridPane.columnIndex="1" GridPane.rowIndex="2">
-
-            <DateTimePicker fx:id="enabledDatePicker" prefHeight="25" prefWidth="280">
-                <tooltip>
-                    <Tooltip text="%dateTimePickerTooltip" />
-                </tooltip>
-            </DateTimePicker>
-            <ComboBox fx:id="relativeDate" prefHeight="25" prefWidth="200" HBox.hgrow="ALWAYS">
-                <tooltip>
-                    <Tooltip text="%relativeDateTooltip" />
-                </tooltip>
-            </ComboBox>
-        </HBox>
-
-        <!-- Row 3 - table views for guidance, displays... -->
-        <fx:include fx:id="optionsTablesView" source="OptionsTables.fxml" prefHeight="800" prefWidth="400" GridPane.columnIndex="0" GridPane.rowIndex="3" GridPane.columnSpan="2" GridPane.hgrow="ALWAYS"/>
-
-        <rowConstraints>
-            <RowConstraints />
-            <RowConstraints />
-            <RowConstraints />
-            <RowConstraints />
-        </rowConstraints>
-
-    </GridPane>
-</ScrollPane>
