+++ app/alarm/ui/src/main/resources/org/phoebus/applications/alarm/ui/config/LeafConfigDialog.fxml
@@ -52,46 +52,29 @@
             </CheckBox>
         </HBox>
 
-        <!-- Row 3 – Disable until (DateTimePicker + relative ComboBox) -->
-        <Label fx:id="disableUntilLabel" text="%disableUntil" GridPane.columnIndex="0" GridPane.rowIndex="3" />
-
-        <HBox fx:id="untilBox" alignment="CENTER" spacing="10" GridPane.columnIndex="1" GridPane.rowIndex="3">
-
-            <DateTimePicker fx:id="enabledDatePicker" prefHeight="25" prefWidth="280">
-                <tooltip>
-                    <Tooltip text="%dateTimePickerTooltip" />
-                </tooltip>
-            </DateTimePicker>
-            <ComboBox fx:id="relativeDate" prefHeight="25" prefWidth="200" HBox.hgrow="ALWAYS">
-                <tooltip>
-                    <Tooltip text="%relativeDateTooltip" />
-                </tooltip>
-            </ComboBox>
-        </HBox>
-
-        <!-- Row 4 – Alarm Delay -->
-        <Label fx:id="delayLabel" text="%alarmDelay" GridPane.columnIndex="0" GridPane.rowIndex="4" />
+        <!-- Row 3 – Alarm Delay -->
+        <Label fx:id="delayLabel" text="%alarmDelay" GridPane.columnIndex="0" GridPane.rowIndex="3" />
 
         <Spinner fx:id="delay" editable="true" prefWidth="80" GridPane.columnIndex="1" GridPane.rowIndex="4" />
 
-        <!-- Row 5 – Alarm Count -->
-        <Label fx:id="countLabel" text="%alarmCount" GridPane.columnIndex="0" GridPane.rowIndex="5" />
-        <Spinner fx:id="count" editable="true" prefWidth="80" GridPane.columnIndex="1" GridPane.rowIndex="5">
+        <!-- Row 4 – Alarm Count -->
+        <Label fx:id="countLabel" text="%alarmCount" GridPane.columnIndex="0" GridPane.rowIndex="4" />
+        <Spinner fx:id="count" editable="true" prefWidth="80" GridPane.columnIndex="1" GridPane.rowIndex="4">
             <tooltip>
                 <Tooltip text="%alarmCountTooltip" />
             </tooltip>
         </Spinner>
 
-        <!-- Row 6 – Enabling Filter -->
-        <Label fx:id="filterLabel" text="%enablingFilter" GridPane.columnIndex="0" GridPane.rowIndex="6" />
-        <TextField fx:id="filter" GridPane.columnIndex="1" GridPane.hgrow="ALWAYS" GridPane.rowIndex="6">
+        <!-- Row 5 – Enabling Filter -->
+        <Label fx:id="filterLabel" text="%enablingFilter" GridPane.columnIndex="0" GridPane.rowIndex="5" />
+        <TextField fx:id="filter" GridPane.columnIndex="1" GridPane.hgrow="ALWAYS" GridPane.rowIndex="5">
             <tooltip>
                 <Tooltip text="%enablingFilterTooltip" />
             </tooltip>
         </TextField>
 
-        <!-- Row 7 - table views for guidance, displays... -->
-        <fx:include fx:id="optionsTablesView" source="OptionsTables.fxml" prefHeight="800" prefWidth="400" GridPane.columnIndex="0" GridPane.rowIndex="7" GridPane.columnSpan="2" GridPane.hgrow="ALWAYS"/>
+        <!-- Row 6 - table views for guidance, displays... -->
+        <fx:include fx:id="optionsTablesView" source="OptionsTables.fxml" prefHeight="800" prefWidth="400" GridPane.columnIndex="0" GridPane.rowIndex="6" GridPane.columnSpan="2" GridPane.hgrow="ALWAYS"/>
 
         <rowConstraints>
             <RowConstraints />
